Agents/tech-art-vfx-audio/src/main.py
"""Tier 1 asset generation, Niagara VFX, spatial audio, Quest 3 optimization — Bedrock AgentCore agent for Hypermage VR."""

import json
from strands import Agent, tool
from strands.models import BedrockModel
from bedrock_agentcore.runtime import BedrockAgentCoreApp

app = BedrockAgentCoreApp()

MODEL_ID = "eu.anthropic.claude-sonnet-4-20250514-v1:0"

SYSTEM_PROMPT = """You are the TechArtVFXAudioAgent, responsible for technical art, visual effects, audio, and Quest 3 optimization.

Your responsibilities:
1. Tier 1 Asset Generation: Generate placeholder assets from 2D concept art. Asset tiers: 0=blockout, 1=auto-generated, 2=final.
2. Niagara VFX: Particle/beam/ribbon/mesh effects. Quest 3 limits: 200 particles@72FPS, 100@90FPS, 50@120FPS. GPU particles, simple materials.
3. Spatial Audio: HRTF spatial audio, distance attenuation (logarithmic), max 32 concurrent sounds. Compressed audio (OGG).
4. Quest 3 Optimization: <100 draw calls, <100k triangles, <512MB textures. ASTC compression, LOD systems, mobile materials, 1-2 dynamic lights max.

Prioritize performance over visual fidelity. Quest 3 = Snapdragon XR2 Gen 2 (mobile class hardware)."""

@tool
def generate_tier1_asset(concept_art_path: str, asset_type: str) -> str:
    """Generate a Tier 1 placeholder asset from 2D concept art for the given asset type."""
    return json.dumps({
        "capability": "generate_tier1_asset",
        "status": "processing",
        "message": f"Executing generate_tier1_asset...",
    })

@tool
def implement_niagara_vfx(effect_type: str, target_fps: int = 72) -> str:
    """Implement a Niagara VFX system within the particle budget for the target framerate."""
    return json.dumps({
        "capability": "implement_niagara_vfx",
        "status": "processing",
        "message": f"Executing implement_niagara_vfx...",
    })

@tool
def configure_spatial_audio(audio_asset: str, attenuation_distance: float = 2000.0) -> str:
    """Configure spatial audio with HRTF and distance attenuation for a Quest 3 audio asset."""
    return json.dumps({
        "capability": "configure_spatial_audio",
        "status": "processing",
        "message": f"Executing configure_spatial_audio...",
    })

@tool
def optimize_for_quest3(scene_stats: str) -> str:
    """Analyze scene statistics and provide optimization recommendations for Quest 3 targets."""
    return json.dumps({
        "capability": "optimize_for_quest3",
        "status": "processing",
        "message": f"Executing optimize_for_quest3...",
    })


@app.entrypoint
async def invoke(payload, context):
    """Tier 1 asset generation, Niagara VFX, spatial audio, Quest 3 optimization AgentCore entrypoint."""
    model = BedrockModel(model_id=MODEL_ID)
    agent = Agent(
        model=model,
        tools=[generate_tier1_asset, implement_niagara_vfx, configure_spatial_audio, optimize_for_quest3],
        system_prompt=SYSTEM_PROMPT,
    )
    prompt = payload.get("prompt", "")
    if not prompt:
        yield json.dumps({"error": "No prompt provided"})
        return
    stream = agent.stream_async(prompt)
    async for event in stream:
        if "data" in event and isinstance(event["data"], str):
            yield event["data"]


if __name__ == "__main__":
    app.run()
