Agents/tech-art-vfx-audio/src/requirements.txt
strands-agents>=1.13.0
bedrock-agentcore>=1.0.3
mcp>=1.19.0
strands-agents-tools>=0.2.16
